Switched newly created dropdown to native dropdown
