+++ projects/freyja-ui/src/lib/card-item/card-item.component.html
@@ -89,7 +89,7 @@
         >
           <fj-button class="card-item__actions__item" color="light" size="small" *ngIf="options?.secondaryAction" (click)="onActionClicked($event, options?.secondaryAction.value)">{{options?.secondaryAction.label}}</fj-button>
           <fj-button class="card-item__actions__item" *ngIf="options?.primaryAction" size="small" (click)="onActionClicked($event, options?.primaryAction.value)">{{options?.primaryAction.label}}</fj-button>
-          <fj-native-dropdown class="card-item__actions__item card-item__dropdown" *ngIf="options?.otherActions" size="small" [options]="options?.otherActions" [customStyles]="{'width' : '40px'}" [showDefaultOnCLick]="true" (optionSelected)="onActionClicked($event)"></fj-native-dropdown>
+          <fj-native-dropdown class="card-item__actions__item card-item__dropdown" *ngIf="options?.otherActions" size="small" [options]="options?.otherActions" [customStyles]="{'width' : '40px'}" [showDefaultOnClick]="true" (optionSelected)="onActionClicked($event)"></fj-native-dropdown>
         </div>
       </ng-container>
 
@@ -168,7 +168,7 @@
     >
       {{options?.belowMenuPrimaryAction.label}}
     </fj-button>
-    <fj-native-dropdown class="card-item__dropdown" *ngIf="showBelowMenu && options?.belowMenuOtherActions" [options]="options?.belowMenuOtherActions" [customStyles]="{'width' : '50px'}" [showDefaultOnCLick]="true" (optionSelected)="onActionClicked($event)"></fj-native-dropdown>
+    <fj-native-dropdown class="card-item__dropdown" *ngIf="showBelowMenu && options?.belowMenuOtherActions" [options]="options?.belowMenuOtherActions" [customStyles]="{'width' : '50px'}" [showDefaultOnClick]="true" (optionSelected)="onActionClicked($event)"></fj-native-dropdown>
   </div>
 </div>
