--- projects/freyja-ui/src/lib/custom-dropdown/custom-dropdown-theme.scss
@@ -1,46 +1,0 @@
-@mixin customDropdownColorsTheme($theme-colors) {
-  .fj-custom-dropdown {
-    select {
-      border: rem-calc(1) solid map-get($theme-colors, "border");
-      color: map-get($theme-colors, "text-header");
-
-      &:focus {
-        border-color: map-get($theme-colors, "color-primary");
-      }
-
-      &:hover {
-        color: map-get($theme-colors, "text-header");
-      }
-
-      &:disabled {
-        background-color: map-get($theme-colors, "color-disable");
-        border-color: map-get($theme-colors, "color-disable");
-
-        &::placeholder {
-          color: map-get($theme-colors, "color-input");
-        }
-      }
-    }
-
-    &--error {
-      select {
-        border-color: map-get($theme-colors, "color-danger");
-      }
-    }
-  }
-
-  label {
-    &.label--disabled {
-      color: map-get($theme-colors, "input");
-    }
-  }
-
-  .input-error {
-    color: map-get($theme-colors, "color-danger");
-    border-color: map-get($theme-colors, "color-danger");
-  }
-
-  .error--label {
-    color: map-get($theme-colors, "color-danger");
-  }
-}
