--- projects/freyja-ui/src/lib/custom-dropdown/custom-dropdown.component.html
@@ -1,49 +1,0 @@
-<!-- With reactive forms -->
-<div
-  class="fj-custom-dropdown"
-  [ngClass]="generateClassList()"
-  *ngIf="formGroup"
-  [formGroup]="formGroup"
->
-  <label
-    [attr.for]="id"
-    class="label--medium"
-    *ngIf="label"
-    >{{label}}</label
-  >
-  <select
-    [id]="id"
-    [tabIndex]="tabIndex"
-    [disabled]="disabled"
-    [attr.name]="name"
-    (click)="onClick($event)"
-    (change)="onOptionSelected($event)"
-    (focus)="onFocus($event)"
-    formControlName="{{formControlName}}"
-  >
-    <option *ngIf="placeholder" [ngValue]="''" hidden>{{placeholder}}</option>
-    <option *ngFor="let option of options" [ngValue]="option.value">{{option.label}}</option>
-  </select>
-  <label class="error--label" *ngIf="error">{{error}}</label>
-</div>
-
-<!-- Without reactive forms -->
-<div
-  class="fj-custom-dropdown"
-  [ngClass]="generateClassList()" *ngIf="!formGroup">
-  <label [attr.for]="id" class="label--medium" *ngIf="label">{{label}}</label>
-  <select
-    [id]="id"
-    [tabIndex]="tabIndex"
-    [disabled]="disabled"
-    [attr.name]="name"
-    (click)="onClick($event)"
-    (change)="onOptionSelected($event)"
-    (focus)="onFocus($event)"
-    [ngStyle]="customStyles"
-  >
-    <option *ngIf="placeholder" hidden>{{placeholder}}</option>
-    <option *ngFor="let option of options" value="{{option.value}}">{{option.label}}</option>
-  </select>
-  <label class="error--label" *ngIf="error">{{error}}</label>
-</div>
