+++ projects/freyja-ui/src/lib/native-dropdown/native-dropdown.component.html
@@ -1,34 +1,49 @@
-<div [formGroup]="formGroup" *ngIf="formGroup">
+<!-- With reactive forms -->
+<div
+  class="fj-native-dropdown"
+  [ngClass]="generateClassList()"
+  *ngIf="formGroup"
+  [formGroup]="formGroup"
+>
   <label
     [attr.for]="id"
     class="label--medium"
     *ngIf="label"
-  >{{label}}</label>
-  <select class="fj-native-dropdown"
-          [ngClass]="generateClassList()"
-          id="{{id}}"
-          (click)="onDropdownClicked($event)"
-          (change)="onOptionSelected($event)"
-          [ngStyle]="customStyles"
-          formControlName="{{controlName}}"
+  >{{label}}</label
   >
+  <select
+    [id]="id"
+    [tabIndex]="tabIndex"
+    [disabled]="disabled"
+    [attr.name]="name"
+    (click)="onClick($event)"
+    (change)="onOptionSelected($event)"
+    (focus)="onFocus($event)"
+    formControlName="{{formControlName}}"
+  >
+    <option *ngIf="placeholder" [ngValue]="''" hidden>{{placeholder}}</option>
     <option *ngFor="let option of options" [ngValue]="option.value">{{option.label}}</option>
   </select>
+  <label class="error--label" *ngIf="error">{{error}}</label>
 </div>
-<!-- Input without reactive forms-->
-<div *ngIf="!formGroup">
-  <label
-    [attr.for]="id"
-    class="label--medium"
-    *ngIf="label"
-  >{{label}}</label>
-  <select class="fj-native-dropdown"
-          id="{{id}}"
-          [ngClass]="generateClassList()"
-          (click)="onDropdownClicked($event)"
-          (change)="onOptionSelected($event)"
-          [ngStyle]="customStyles"
+
+<!-- Without reactive forms -->
+<div
+  class="fj-native-dropdown"
+  [ngClass]="generateClassList()" *ngIf="!formGroup">
+  <label [attr.for]="id" class="label--medium" *ngIf="label">{{label}}</label>
+  <select
+    [id]="id"
+    [tabIndex]="tabIndex"
+    [disabled]="disabled"
+    [attr.name]="name"
+    (click)="onClick($event)"
+    (change)="onOptionSelected($event)"
+    (focus)="onFocus($event)"
+    [ngStyle]="customStyles"
   >
+    <option *ngIf="placeholder" hidden>{{placeholder}}</option>
     <option *ngFor="let option of options" value="{{option.value}}">{{option.label}}</option>
   </select>
+  <label class="error--label" *ngIf="error">{{error}}</label>
 </div>
