+++ projects/freyja-ui/src/lib/native-dropdown/native-dropdown.component.scss
@@ -1,55 +1,102 @@
 @import "../core/_config";
 
 .fj-native-dropdown {
-  /* remove arrows for Firefox */
-  -moz-appearance: none;
-  /* remove arrows for for Chrome */
-  -webkit-appearance: none;
-  appearance: none;
-  font-family: OpenSans, sans-serif;
-  font-size: 1rem;
-  height: rem-calc(40);
-  padding: rem-calc(0 15);
-  cursor: pointer;
-  border: rem-calc(1) solid $color-blue;
-  border-radius: rem-calc(4);
-  /* remove arrows for IE */
-  &::-ms-expand {
-    display: none;
+  select {
+    display: block;
+    -moz-appearance: none;
+    -webkit-appearance: none;
+    align-items: center;
+    vertical-align: middle;
+    outline: none;
+    height: rem-calc(40);
+    margin-top: .25em;
+    font-size: .95em;
+    box-sizing: border-box;
+    border-radius: rem-calc(4);
+    background-image: url("data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABAAAAAUCAMAAACzvE1FAAAADFBMVEUzMzMzMzMzMzMzMzMKAG/3AAAAA3RSTlMAf4C/aSLHAAAAPElEQVR42q3NMQ4AIAgEQTn//2cLdRKppSGzBYwzVXvznNWs8C58CiussPJj8h6NwgorrKRdTvuV9v16Afn0AYFOB7aYAAAAAElFTkSuQmCC");
+    background-size: rem-calc(8 10);
+    background-repeat: no-repeat;
+    background-position: right rem-calc(9) center;
+    background-image: none \9; // IE 7, 8, & 9
+    transition: .2s linear;
+
+    // remove arrows for IE
+    &::-ms-expand {
+      display: none;
+    }
+
+    &:disabled {
+      cursor: not-allowed;
+    }
+  }
+
+  &--full-width {
+    select {
+      width: 100%;
+      display: flex;
+      justify-content: center;
+    }
   }
 
   /* sizing variants */
   &--large {
-    font-size: $font-base;
-    height: rem-calc(48);
-    line-height: $font-base-line-height;
-    padding: rem-calc(14 20);
+    select {
+      font-size: $font-base;
+      height: rem-calc(48);
+      line-height: $font-base-line-height;
+      padding: rem-calc(0 40 0 20);
+    }
   }
+
   &--medium {
-    font-size: $font-small;
-    height: rem-calc(40);
-    line-height: $font-small-line-height;
-    padding: rem-calc(11 16);
+    select {
+      font-size: $font-small;
+      height: rem-calc(40);
+      line-height: $font-small-line-height;
+      padding: rem-calc(0 32 0 16);
+    }
   }
+
   &--small {
-    font-size: $font-extra-small;
-    height: rem-calc(32);
-    line-height: $font-extra-small-line-height;
-    padding: rem-calc(8 12)
+    select {
+      font-size: $font-extra-small;
+      height: rem-calc(32);
+      line-height: $font-extra-small-line-height;
+      padding: rem-calc(0 24 0 12);
+    }
   }
+
   &--micro {
-    font-size: $font-label;
-    height: rem-calc(24);
-    line-height: $font-label-line-height;
-    padding: rem-calc(6 8);
+    select {
+      font-size: $font-label;
+      height: rem-calc(24);
+      line-height: $font-label-line-height;
+      padding: rem-calc(0 20 0 8);
+    }
+  }
+
+  &--required {
+    label:after {
+      content: " *";
+    }
   }
 }
 
 label {
   cursor: pointer;
   font-size: 0.875rem;
+
   &.label--medium {
     font-size: .75rem;
-    padding-left: .5em;
   }
+}
+
+.input-error {
+  font-size: .75rem;
+  padding: .5em;
+}
+
+.error--label {
+  font-size: rem-calc(12);
+  padding-left: .5em;
 }
