+++ projects/freyja-ui/src/lib/native-dropdown/native-dropdown.component.ts
@@ -1,7 +1,11 @@
-import { Component, EventEmitter, Input, Output } from '@angular/core';
-import {BaseComponent} from '../shared/base.component';
-import {FormGroup} from '@angular/forms';
+import { AfterViewInit, Component, EventEmitter, Input, OnChanges, Output } from '@angular/core';
+import { FormGroup } from '@angular/forms';
 
+import { BaseComponent } from '@libs/freyja-ui/src/lib/shared/base.component';
+
+/**
+ * Count for next id
+ */
 let nextId = 0;
 
 @Component({
@@ -9,63 +13,128 @@ let nextId = 0;
   templateUrl: './native-dropdown.component.html',
   styleUrls: ['./native-dropdown.component.scss']
 })
-export class NativeDropdownComponent extends BaseComponent {
+export class NativeDropdownComponent extends BaseComponent implements AfterViewInit, OnChanges {
   /**
-   * options of the dropdown
+   * Dropdown options
+   * Each option be an object with properties label and value
+   * {label: string, value: number | string }
    */
   @Input() options = [];
+
   /**
    * dropdown id
    */
-  @Input() id = `fj-native-dd-${++nextId}`;
+  @Input() id = `fj-native-dropdown-${++nextId}`;
+
   /**
-   * dropdown width
+   * dropdown with
    */
   @Input() width: string;
+
   /**
-   * wether or not to show default value after selecting an option
+   *  Form group to which the control belongs to
    */
-  @Input() showDefaultOnCLick = false;
+  @Input() formGroup: FormGroup;
+
   /**
-   * event that is triggered when an option is selected
+   * Form control name for reactive forms
    */
-  @Output() optionSelected: EventEmitter<any> = new EventEmitter();
+  @Input() formControlName: string;
+
   /**
    * Error Message to be displayed
    */
   @Input() error: string;
+
   /**
-   * reactive form control name
+   * To Apply forceful error state to input component
    */
-  @Input() controlName: string;
+  @Input() errored = false;
+
   /**
-   * reactive form group
+   * checks if the element should be focused by default
    */
-  @Input() formGroup: FormGroup;
+  @Input() focused = false;
+
+  /**
+   * Size can be micro, small, medium or large
+   */
+  @Input() size: 'micro' | 'small' | 'medium' | 'large' = 'medium';
+
+  /**
+   * custom styles for dropdown
+   */
+  @Input() customStyles;
+
+  /**
+   * whether or not to show default value after selecting an option
+   */
+  @Input() showDefaultOnClick = false;
+
+  /**
+   * event that is triggered when an option is selected
+   * called with value of dropdown
+   */
+  @Output() optionSelected: EventEmitter<any> = new EventEmitter();
+
+  /**
+   * Callback for value change
+   */
+  @Output() change: EventEmitter<any> = new EventEmitter<any>();
+
+  /**
+   * Callback for focus event
+   */
+  @Output() focus: EventEmitter<any> = new EventEmitter<any>();
+
+  /**
+   * Callback for click event
+   */
+  @Output() click: EventEmitter<any> = new EventEmitter<any>();
 
   constructor() {
     super();
   }
 
-  /**
-   * function triggered when an option is selected
-   * @param evt - mouse click event
-   */
-  onOptionSelected(evt) {
-    evt.stopPropagation();
-    const dropdownElem: any = document.getElementById(this.id);
-    this.optionSelected.emit(dropdownElem.value);
-    if (this.showDefaultOnCLick) {
-      // reselect default value
-      dropdownElem.value = this.options[0].label;
+  ngOnChanges() {
+    const dropdownElement = document.getElementById(this.id);
+    if (this.focused && dropdownElement) {
+      dropdownElement.focus();
+    }
+  }
+
+  ngAfterViewInit(): void {
+    const dropdownElement = document.getElementById(this.id);
+    if (this.focused && dropdownElement) {
+      dropdownElement.focus();
+    }
+  }
+
+  onOptionSelected(event: Event) {
+    event.stopPropagation();
+    const dropdownElement: any = document.getElementById(this.id);
+    this.optionSelected.emit(dropdownElement.value);
+
+    if (this.showDefaultOnClick) {
+      dropdownElement.value = this.options[0].label;
     }
   }
 
   /**
-   * @param evt - mouse click event
+   * triggers click event
+   * @param event HTML click event
    */
-  onDropdownClicked(evt) {
-    evt.stopPropagation();
+  onClick(event: Event) {
+    event.stopPropagation();
+    this.click.emit(event);
+  }
+
+  /**
+   * triggers focus event
+   * @param event HTML focus event
+   */
+  onFocus(event: Event) {
+    this.focus.emit(event);
   }
 
   /**
@@ -75,8 +144,10 @@ export class NativeDropdownComponent extends BaseComponent {
     const baseCls = 'fj-native-dropdown--';
     const classes = {
       [baseCls + this.size]: this.size,
+      [baseCls + 'required']: this.required,
+      [baseCls + 'error']: this.errored || this.error,
+      [baseCls + 'full-width']: this.fullWidth
     };
     return classes;
   }
-
 }
