--- src/app/pages/custom-dropdown/custom-dropdown.component.html
@@ -1,38 +1,0 @@
-<fj-demo-content-wrap>
-  <h2>Custom Dropdown</h2>
-  <h3>Basic custom dropdown</h3>
-  <div class="dropdown-wrapper">
-    <fj-custom-dropdown [options]="options"></fj-custom-dropdown>
-  </div>
-  <h3>Custom dropdown with label</h3>
-  <fj-custom-dropdown [options]="options" label="Label"></fj-custom-dropdown>
-  <fj-custom-dropdown [options]="options" label="Label" [required]="true"></fj-custom-dropdown>
-
-  <h3>Size</h3>
-  <div>A custom dropdown can have different sizes.</div>
-  <div class="dropdown-wrapper">
-    <fj-custom-dropdown [options]="options" size="large" label="Large"></fj-custom-dropdown>
-    <fj-custom-dropdown [options]="options" size="medium" label="Medium"></fj-custom-dropdown>
-    <fj-custom-dropdown [options]="options" size="small" label="Small"></fj-custom-dropdown>
-    <fj-custom-dropdown [options]="options" size="micro" label="Micro"></fj-custom-dropdown>
-  </div>
-  <h3>Full width</h3>
-  <fj-custom-dropdown [options]="options" [fullWidth]="true"></fj-custom-dropdown>
-  <h3>Disabled</h3>
-  <fj-custom-dropdown [options]="options" disabled></fj-custom-dropdown>
-  <h3>State</h3>
-  <fj-custom-dropdown [options]="options" label="Error Example" error="Example error message"></fj-custom-dropdown>
-  <h3>Placeholder</h3>
-  <fj-custom-dropdown [options]="options" label="Placeholder" placeholder="placeholder Example"></fj-custom-dropdown>
-  <h3>Focused by default</h3>
-  <fj-custom-dropdown [options]="options" label="Focused" [focused]="true"></fj-custom-dropdown>
-  <h3>Reactive Forms</h3>
-  <fj-custom-dropdown [options]="options"
-                      label="Test FormGroup"
-                      placeholder="Choose option"
-                      [formGroup]="testFormGroup"
-                      formControlName="selectValue"
-                      (optionSelected)="onOptionSelected($event)"
-  ></fj-custom-dropdown>
-  <h6>Selected Value: {{testFormGroup?.value?.selectValue}}</h6>
-</fj-demo-content-wrap>
