+++ src/app/pages/native-dropdown/native-dropdown.component.scss
@@ -1,0 +1,9 @@
+.dropdown-wrapper {
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+
+  fj-custom-dropdown {
+    margin-right: .5rem;
+  }
+}
