+++ src/app/pages/native-dropdown/native-dropdown.component.ts
@@ -1,5 +1,6 @@
 import {Component, OnInit} from '@angular/core';
 import {TextHelper} from '@freyjaDemo/app/shared/helpers/text.helper';
+import { FormControl, FormGroup } from '@angular/forms';
 
 @Component({
   selector: 'fj-demo-native-dropdown',
@@ -7,6 +8,7 @@ import {TextHelper} from '@freyjaDemo/app/shared/helpers/text.helper';
   styleUrls: ['./native-dropdown.component.scss']
 })
 export class DemoNativeDropdownComponent implements OnInit {
+
   public options = [
     {
       label: 'Option 1',
@@ -26,12 +28,23 @@ export class DemoNativeDropdownComponent implements OnInit {
     }
   ];
 
-  public nativeDropdown = this.textHelper.dedent(`<fj-native-dropdown [options]="options" (optionSelected)="onOptionSelected($event)">
-                           </fj-native-dropdown>`);
+  testFormGroup: FormGroup;
+
+  public sizeShowCase = this.textHelper.dedent(`
+    <fj-custom-dropdown [options]="options" size="large"></fj-custom-dropdown>
+    <fj-custom-dropdown [options]="options" size="medium"></fj-custom-dropdown>
+    <fj-custom-dropdown [options]="options" size="small"></fj-custom-dropdown>
+    <fj-custom-dropdown [options]="options" size="micro"></fj-custom-dropdown>`);
+
   constructor(private textHelper: TextHelper) {}
 
   ngOnInit() {
+    this.testFormGroup = new FormGroup({
+      selectValue: new FormControl('')
+    });
   }
 
-  onOptionSelected($event) {}
+  onOptionSelected(event) {
+    console.log(this.testFormGroup.value);
+  }
 }
